+++ fota/fota_platform.h
@@ -21,7 +21,7 @@
 
 #include "fota/fota_base.h"
 
-#if MBED_CLOUD_CLIENT_FOTA_ENABLE
+#if defined(MBED_CLOUD_CLIENT_FOTA_ENABLE)
 
 #include "fota/fota_status.h"
 #include "fota/fota_block_device.h"
@@ -98,5 +98,5 @@ static inline int fota_platform_abort_update_hook(const char *comp_name)
 }
 #endif
 
-#endif // MBED_CLOUD_CLIENT_FOTA_ENABLE
+#endif // defined(MBED_CLOUD_CLIENT_FOTA_ENABLE)
 #endif // __FOTA_PLATFORM_H_