+++ mbed-client/source/m2mnsdlinterface.cpp
@@ -85,7 +85,7 @@
 #define TRACE_GROUP "mClt"
 #define MAX_QUERY_COUNT 10
 
-const char *MCC_VERSION = "mccv=4.7.0";
+const char *MCC_VERSION = "mccv=4.7.1-multicast";
 
 int8_t M2MNsdlInterface::_tasklet_id = -1;